feat / lookitem

+++ FashionToday_Back/src/main/java/hamburger/fashiontoday/domain/lookitemclass/LookItemClassRepository.java
@@ -3,5 +3,5 @@ package hamburger.fashiontoday.domain.lookitemclass;
 import hamburger.fashiontoday.domain.member.Member;
 import org.springframework.data.repository.CrudRepository;
 
-public interface LookItemClassRepository extends CrudRepository<LookitemClass,Long> {
+public interface LookItemClassRepository extends CrudRepository<LookitemClass,Integer> {
 }
